[Misc] Clean up more utils (#27567)
