+++ requirements/docs.txt
@@ -13,6 +13,8 @@ ruff
 # Required for argparse hook only
 -f https://download.pytorch.org/whl/cpu
 cachetools
+cloudpickle
+py-cpuinfo
 msgspec
 pydantic
 torch
